requirements/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
amqp==2.6.1               # via kombu
beautifulsoup4==4.6.0     # via wagtail
billiard==3.6.3.0         # via celery
celery==4.4.7             # via -r requirements.in
certifi==2020.6.20        # via requests, sentry-sdk
chardet==3.0.4            # via requests
click==7.1.2              # via geobuf
django-modelcluster==5.1  # via wagtail
django-modeltranslation==0.16  # via wagtail-modeltranslation
django-rest-framework-condition==0.1.1  # via -r requirements.in
django-reversion==3.0.8   # via -r requirements.in
django-rosetta==0.9.5     # via -r requirements.in
django-sendfile==0.3.11   # via -r requirements.in
django-taggit==1.3.0      # via wagtail
django-treebeard==4.3.1   # via wagtail
django==2.2.17            # via -r requirements.in, django-modeltranslation, django-rest-framework-condition, django-reversion, django-rosetta, django-sendfile, django-taggit, django-treebeard, djangorestframework, wagtail
djangorestframework-gis==0.16  # via -r requirements.in
djangorestframework-jwt==1.11.0  # via -r requirements.in
djangorestframework==3.12.1  # via djangorestframework-gis, wagtail
draftjs-exporter==2.1.7   # via wagtail
geobuf==1.1.1             # via -r requirements.in
html5lib==1.1             # via wagtail
idna==2.10                # via requests
importlib-resources==3.3.0  # via -r requirements.in
kombu==4.6.11             # via celery
pillow==7.2.0             # via wagtail
polib==1.1.0              # via django-rosetta
protobuf==3.13.0          # via -r requirements.in, geobuf
psycopg2-binary==2.8.6    # via -r requirements.in
pyjwt==1.7.1              # via djangorestframework-jwt
pytz==2020.4              # via celery, django, django-modelcluster, wagtail
requests==2.24.0          # via django-rosetta, wagtail
sentry-sdk==0.19.2        # via -r requirements.in
six==1.15.0               # via django-modeltranslation, django-rosetta, geobuf, html5lib, protobuf, wagtail
sqlparse==0.4.1           # via django
unidecode==1.1.1          # via wagtail
urllib3==1.25.11          # via requests, sentry-sdk
vine==1.3.0               # via amqp, celery
wagtail-modeltranslation==0.10.15  # via -r requirements.in
wagtail==2.7.4            # via -r requirements.in, wagtail-modeltranslation
webencodings==0.5.1       # via html5lib
willow==1.3               # via wagtail
xlrd==1.2.0               # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
